% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_from_db.R
\name{annot_from_db}
\alias{annot_from_db}
\title{Annotate from Database}
\usage{
annot_from_db(peaks_matched, dir_mgf, fname_regex)
}
\arguments{
\item{peaks_matched}{Tibble with column 'spectrum' containing MSnBase Spectrum objects encapsulated in lists}

\item{dir_mgf}{Directory containing MS1 spectra (.mgf format, suffix) annotated using TITLE= tag}

\item{fname_regex}{Regex string that can be used to prefilter the TITLE= tags}
}
\description{
Simple iterative database search to annotate spectra in a tibble
}
\examples{
# effectively add an annotation column to peaks_matched:
peaks_matched <- annot_from_db(peaks_matched, "~/msdb/", "methyl")
}
\keyword{annot}
\keyword{annotate}
\keyword{database}
\keyword{db}
