% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_from_db2.R
\name{annot_from_db2}
\alias{annot_from_db2}
\title{Annotate from Database}
\usage{
annot_from_db2(spectra, dir_db, bin, fname_regex = "*.mzXML", cores = 1)
}
\arguments{
\item{spectra}{Grouped (e.g. by \code{scan} tibble with columns \code{mz}, \code{intensity}.}

\item{dir_db}{Directory containing database spectrum files in an mzR-readable format (mzXML, mzML, etc.)}

\item{fname_regex}{Regex string that can be used to prefilter database files (generally the suffix.)}
}
\value{
Tibble of cosine scores for each spectrum against each file
}
\description{
Simple iterative database search to annotate spectra in a tibble
}
\examples{
# effectively add an annotation column to peaks_matched:
peaks_matched <- annot_from_db(peaks_matched, "~/msdb/", "methyl")
}
\keyword{annot}
\keyword{annotate}
\keyword{database}
\keyword{db}
