% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve (AUC)}
\usage{
auc(chromdata)
}
\arguments{
\item{chromdata}{Tibble with columns \code{rt} and \code{intensity}, pre-grouped by peak (e.g. by ROI.)}
}
\value{
A summary of the input tibble with additional columns:
\code{rtmin}, \code{rtmax},
\code{into} (integral overall),
\code{intb}, (integral baseline-corrected),
}
\description{
Summarize grouped chromatographic data (windowed chromatograms),
returning the RT window as well as overall and baseline-corrected integrals.
}
\examples{

areas_rois <- xics \%>\%
 group_by(roi) \%>\%
 auc()
}
\keyword{area}
\keyword{curve}
\keyword{integral}
\keyword{peak}
