% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_runs.R
\name{average_runs}
\alias{average_runs}
\title{Average Chromatographic Runs}
\usage{
average_runs(exp, samp_indices, bin)
}
\arguments{
\item{exp}{MSNExp object; all samples to be averaged need to be in the experiment.}

\item{samp_indices}{Vector of sample indices to be averaged.}

\item{bin}{Bin spectra to this m/z tolerance. If missing, do not bin at all.}
}
\description{
Take an MSNExp and a vector of sample indices. Average those runs,
save the resultant CDF, and return that filename.
}
\examples{
fname_master_blank <- average_runs(exp, samp_indices = c(1,2,3))

}
\keyword{average}
\keyword{blank}
\keyword{experiment}
\keyword{run}
