% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_spectra.R
\name{average_spectra}
\alias{average_spectra}
\title{Average Spectra}
\usage{
average_spectra(specs, bin)
}
\arguments{
\item{specs}{A list of spectra}

\item{bin}{Bin spectra to this m/z tolerance. If missing, do not bin at all.}
}
\description{
Take a list of MSnbase Spectrum objects and average the intensities, no scaling or anything.
Return another Spectrum. Use e.g. for averaging blank runs.
}
\examples{
spec_avg <- average_spectrum(spec1, spec2)

}
\keyword{baseline}
\keyword{blank}
\keyword{spectra}
\keyword{spectrum}
\keyword{subtract}
