% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_spectra.R
\name{bin_spectra}
\alias{bin_spectra}
\title{Bin Spectra}
\usage{
bin_spectra(chromdata, x = "mz", y = "intensity", scan = "scan", bin_width = 1)
}
\arguments{
\item{chromdata}{Tibble with columns identifying spectrum x-coord (e.g. mz or wl), y-coord (intensity, abs, etc.),
and scan number}

\item{x}{Name of x-coordinate column}

\item{y}{Name of y-coordinate column}

\item{scan}{Name of scan ID column}

\item{bin_width}{Width of x bins}
}
\value{
\code{chromdata} tibble with the x-axis binned as specified
}
\description{
Convenience function takes a list of m/z values and rounds them to the nearest integer
with specified precision. Breaks on the 0.5.
}
\examples{
ions_binned <- ions \%>\% bin_spectra(x = "mz", y = "intensity", scan = "scan")

}
\keyword{bin}
\keyword{mass}
\keyword{spectra}
\keyword{spectrum}
\keyword{wavelength}
