% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldr.r
\name{calc_intb_max}
\alias{calc_intb_max}
\title{Calculate LDR Upper Limit (\code{intb_max})}
\usage{
calc_intb_max(peak_group, key, rsq = 0.95, min_dils = 3)
}
\arguments{
\item{peak_group}{Tibble with columns \code{dil}, and \code{intb},
subsetted to a single compound, e.g. by \code{group_map()}}

\item{key}{compound identifier, passed as a one-element tibble, as by \code{group_map()}}

\item{rsq}{minimum R^2 at which signal is considered linear}

\item{min_dils}{minimum number of dilutions that must be present for a compound}
}
\value{
tibble with max. linear peak area, number of data points, R^2, and compound id
}
\description{
Calculate LDR Upper Limit (\code{intb_max})
}
\examples{
use this function through \code{calc_ldrs}
}
\keyword{dynamic}
\keyword{ldr}
\keyword{linear}
\keyword{range}
