% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldr.r
\name{calc_ldrs}
\alias{calc_ldrs}
\title{Calculate LDRs for all peaks
A convenience wrapper for \code{calc_intb_max()}}
\usage{
calc_ldrs(peaks_matched, cpd_id = "rt_std", rsq = 0.98, min_dils = 3)
}
\arguments{
\item{peaks_matched}{Tibble with columns \code{dil}, and \code{intb}}

\item{cpd_id}{String naming the column to be used as a unique compound ID.
Conservatively, \code{rt_std}. If you trust your dereplication, can be \code{id_db}.}

\item{rsq}{minimum R^2 at which signal is considered linear}

\item{min_dils}{minimum number of dilutions that must be present for a compound}
}
\value{
tibble with max. linear peak area, number of data points, R^2, and compound id
}
\description{
Calculate LDRs for all peaks
A convenience wrapper for \code{calc_intb_max()}
}
\examples{

}
\keyword{dynamic}
\keyword{ldr}
\keyword{linear}
\keyword{range}
