% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_peaks.R
\name{cluster_peaks}
\alias{cluster_peaks}
\title{Cluster peaks
Takes a summary table of peaks with \code{rt_min}/\code{rt_max} columns,
clusters the peaks based on fraction overlap,
returns the same tibble with cluster index.}
\usage{
cluster_peaks(peaks, thres_lap)
}
\arguments{
\item{peaks}{Tibble with \code{rt_min}/\code{rt_max}. If there is >1 row per peak, they should be grouped}

\item{thres_lap}{Fraction of RT overlap required for peaks to cluster (calculated as a fraction of the narrower peak)}
}
\value{
Input tibble but with \code{clust_id} column added
}
\description{
Cluster peaks
Takes a summary table of peaks with \code{rt_min}/\code{rt_max} columns,
clusters the peaks based on fraction overlap,
returns the same tibble with cluster index.
}
\examples{
peaks_clustered <- peaks \%>\%
  cluster_peaks(thres_lap = 0.8)

}
\keyword{cluster}
\keyword{peaks}
