% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spectra.R
\name{cluster_spectra}
\alias{cluster_spectra}
\title{Cluster Spectra
Takes a grouped tibble of scan data, clusters scans based on a cosine threshold,
returns the same tibble with a cluster index.}
\usage{
cluster_spectra(spectra, thres_cos, thres_pks = 2, x = "mz", bin, cores = 1)
}
\arguments{
\item{spectra}{Grouped tibble \code{mz}/\code{wl} and \code{intensity}. Should be grouped by scan.}

\item{thres_cos}{Min allowable cosine score for scans to cluster}

\item{thres_pks}{Min allowable number of matched peaks}

\item{x}{Name of x-axis variable (\code{mz} or \code{wl}, etc.)}
}
\value{
Input tibble with \code{clust_id} column added
}
\description{
Cluster Spectra
Takes a grouped tibble of scan data, clusters scans based on a cosine threshold,
returns the same tibble with a cluster index.
}
\examples{
spectra_clustered <- spectra \%>\%
  group_by(scan, file) \%>\%
  cluster_spectra(thres_cos = 0.95)

}
\keyword{cluster}
\keyword{cosine}
\keyword{spectra}
