% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{cosine}
\alias{cosine}
\title{Cosine Similarity of Two Vectors
Lifted verbatim from Fridolin Wild's LSA package (https://cran.r-project.org/web/packages/lsa/)
Serves here as a helper to cosine_spectra().}
\usage{
cosine(x, y = NULL)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2}
}
\value{
cosine similarity score of the two vector (dbl)
}
\description{
Takes two vector-like objects
}
\examples{
cos <- cosine_spectra(ions \%>\% filter(scan == 114), ions \%>\% filter(scan == 114))
> 0.94

}
\keyword{cosine}
\keyword{similarity}
\keyword{vector}
