% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosines_pairwise.R
\name{cosine_pairwise}
\alias{cosine_pairwise}
\title{Calculate Cosine Scores Pairwise
Takes a grouped tibble of scan data, calculates pairwise cosine similarities}
\usage{
cosine_pairwise(
  spectra,
  x = "mz",
  thres_cos = 0,
  thres_rt = Inf,
  cores = 1,
  bin
)
}
\arguments{
\item{spectra}{Grouped tibble \code{mz}/\code{wl} and \code{intensity}. Should be grouped by scan.}

\item{x}{Name of x-axis variable (\code{mz} or \code{wl}, etc.)}

\item{thres_cos}{Min allowable cosine score for scans to cluster}

\item{thres_rt}{Max allowable RT difference for scans to cluster}
}
\value{
Input tibble but with neighbor and cos columns added
}
\description{
Calculate Cosine Scores Pairwise
Takes a grouped tibble of scan data, calculates pairwise cosine similarities
}
\examples{
spectra_clustered <- spectra \%>\%
  group_by(scan, file) \%>\%
  cluster_spectra()

}
\keyword{cluster}
\keyword{spectra}
