% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_peaks.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Detect Peaks
Take chromatographic data, return summary of peaks above a hard threshold, SNR threshold, or biggest n peaks.
Data should be input as one or more groups of \code{rt}: \code{intensity}, without gaps in RT.
NOTE: improper grouping (e.g. by \code{rt}) will cause this function to return an empty tbl.}
\usage{
detect_peaks(
  chromdata,
  thres_snr = 0,
  thres_intensity = 0,
  n = Inf,
  int_min = 1,
  target_rt
)
}
\arguments{
\item{chromdata}{chromatographic data, preferable single-ion or -wavelength}

\item{thres_snr}{intensity SNR threshold for peak detection}

\item{thres_intensity}{Hard intensity threshold for peak detection}

\item{n}{number of peaks to return}

\item{int_min}{Minimum nonzero intensity, used to reasonably calculate SNR}
}
\value{
Peak rows for each group of the passed tibble, with a coords column containing the peak domain
}
\description{
Detect Peaks
Take chromatographic data, return summary of peaks above a hard threshold, SNR threshold, or biggest n peaks.
Data should be input as one or more groups of \code{rt}: \code{intensity}, without gaps in RT.
NOTE: improper grouping (e.g. by \code{rt}) will cause this function to return an empty tbl.
}
\examples{
peaks <- chromdata \%>\%
get_peaks(thres_snr = 20)


}
\keyword{detect}
\keyword{peaks}
