% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2tbl.R
\name{exp2tbl}
\alias{exp2tbl}
\title{Convert MSnExp to Tibble}
\usage{
exp2tbl(exp, bin)
}
\arguments{
\item{exp}{MSNExp object}

\item{bin}{Bin spectra to this m/z tolerance. If missing, do not bin at all.}
}
\description{
Take an MSNExp and convert it losslessly (but exclusive of metadata)
into a tbl with columns rt, mz, and intensity (the tidychrom-ms format.)
}
\examples{
ions <- exp2tbl(exp)

}
\keyword{convert}
\keyword{experiment}
