% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_disk2mem.R
\name{exp_disk2mem}
\alias{exp_disk2mem}
\title{Convert OnDiskMSnExp to MSNExp}
\usage{
exp_disk2mem(exp_disk)
}
\arguments{
\item{exp}{OnDiskMSNExp object. Be sure to subset as desired beforehand (e.g. filterFile())
or else returned object could take gigs of RAM.}
}
\description{
Take an OnDiskMSNExp and load it into memory.
Specifically, this means taking the files in it and loading a new experiment object.
}
\examples{
fname_master_blank <- average_runs(exp, samp_indices = c(1,2,3))

}
\keyword{disk}
\keyword{experiment}
\keyword{memory}
