% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chromatograms.R
\name{extract_chromatograms}
\alias{extract_chromatograms}
\title{Extract Chromatograms}
\usage{
extract_chromatograms(peaks, chromdata, x = "mz", cores = 1)
}
\arguments{
\item{peaks}{Tibble with columns \code{scan} and \code{mz}/\code{wl}, and other identifiers}

\item{chromdata}{Full-spectrum chromatographic data}

\item{x}{Name of x-axis variable (\code{mz} or \code{wl}, etc.)}
}
\value{
A tibble of XICs or single-wavelength chromatograms for the full peaks (ready for integration!)
Peak ID columns like roi are preserved.
}
\description{
Extract Chromatograms
}
\examples{
xics_matched <- peaks_matched \%>\%
extract_chromatograms(chromdata)

}
\keyword{chromatogram}
\keyword{extract}
