% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature2samp.R
\name{feature2samp}
\alias{feature2samp}
\title{Extract Sample Index from MSNbase Feature String}
\usage{
feature2samp(feature)
}
\arguments{
\item{feature}{string of form "Fxx.Sxxxx"}
}
\description{
Extract Sample Index from MSNbase Feature String
}
\examples{
feature2samp("F01.S0004")
>[1] 1

}
\keyword{extract}
\keyword{feature}
\keyword{samp}
\keyword{sample}
