% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature2scan.R
\name{feature2scan}
\alias{feature2scan}
\title{Extract Sample Index from MSNbase Feature String}
\usage{
feature2scan(feature)
}
\arguments{
\item{feature}{string of form "Fxx.Sxxxx"}
}
\description{
Extract Sample Index from MSNbase Feature String
}
\examples{
feature2scan("F01.S0004")
>[1] 4

}
\keyword{extract}
\keyword{feature}
\keyword{samp}
\keyword{sample}
