% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_listoftbl.R
\name{gather_listoftbl}
\alias{gather_listoftbl}
\title{Gather List of Tibbles (or other dataframe-like type)
Names and index of the list are saved to columns}
\usage{
gather_listoftbl(list_in, key = "key", index = "index", cores = 1)
}
\arguments{
\item{list_in}{A named list of tibbles or another dataframe-like type. Must all have matching columns.}

\item{key}{Name for the tbl column that original list names will end up in.}

\item{index}{Name for the tbl column that original list indices will end up in.}

\item{cores}{Number of cores to use for parallel ops on the input list. Only beneficial in huge tables (1E7s of rows.)}
}
\description{
Gather List of Tibbles (or other dataframe-like type)
Names and index of the list are saved to columns
}
\examples{
ions <- ions <- mzr \%>\%
   spectra() \%>\%
   setNames(., header(mzr)$retentionTime) \%>\%
   gather_listoftbl(., key = "rt", index = "scan")

}
\keyword{gather}
\keyword{list}
\keyword{named}
\keyword{tibble}
