% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldr.R
\name{ldr}
\alias{ldr}
\title{Linear Dynamic Range (LDR)}
\usage{
ldr(
  dildata,
  x = "dil",
  y = "intb",
  rsq = 0.95,
  min_dils = 3,
  force_origin = FALSE
)
}
\arguments{
\item{x}{name of independent variable, e.g. concentration or dilution}

\item{y}{name of dependent variable, e.g. into or intb}

\item{rsq}{minimum R^2 at which signal is considered linear}

\item{min_dils}{minimum number of unique x values (dilutions) that must be present for a group}

\item{force_origin}{Force regression through the origin? Set T for pre-blanked data, otherwise F.}
}
\value{
a named list:
\code{$y_max}: maximum \code{y} for which R^2 >= \code{rsq}
\code{$n_dils}: number of dilutions used in the linreg
\code{$rr}: R^2 of the linreg
# NTS: can put slope and intercept in here too, if need be
}
\description{
Take grouped tibble with x and y colnames, also rsq_min value.
Return same tibble with each group truncated to the linear dynamic range.
}
\examples{
# use inside a dplyr::summarise() call
ldrs_roi <- areas_all \%>\%
  group_by(roi) \%>\%
  ldr(rsq = 0.95, min_dils = 4)

}
\keyword{dynamic}
\keyword{ldr}
\keyword{limits}
\keyword{linearity}
\keyword{range}
