% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol.r
\name{lol}
\alias{lol}
\title{Upper Limit of Linearity (LOL)}
\usage{
lol(x, y, rsq = "0.95", min_dils = 3, force_origin = FALSE)
}
\arguments{
\item{x}{independent variable, e.g. concentration or dilution}

\item{y}{dependent variable, e.g. into or intb}

\item{rsq}{minimum R^2 at which signal is considered linear}

\item{min_dils}{minimum number of unique x values (dilutions) that must be present for a compound}

\item{force_origin}{Force regression through the origin? Set T for pre-blanked data, otherwise F.}
}
\value{
a named list:
\code{$y_max}: maximum \code{y} for which R^2 >= \code{rsq}
\code{$n_dils}: number of dilutions used in the linreg
\code{$rr}: R^2 of the linreg
# NTS: can put slope and intercept in here too, if need be
}
\description{
Take lists of x and y, return the upper limit of linearity.
}
\examples{
# use inside a dplyr::summarise() call
ldrs_roi <- areas_all \%>\%
  group_by(roi) \%>\%
  summarise(intb_max = lol(intb))

}
\keyword{limit}
\keyword{linearity}
\keyword{upper}
