% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_coords.R
\name{melt_coords}
\alias{melt_coords}
\title{Melt chromatographic coordinates}
\usage{
melt_coords(chroms)
}
\arguments{
\item{chroms}{An XChromatograms object.}
}
\description{
Takes RT and intensity coordinates from a list of matching chromatograms (TIC, BPC, XIC etc.)
return a long tibble of RT, intensity, samp
}
\keyword{chrom}
\keyword{chromatogram}
\keyword{coord}
\keyword{coordinates}
\keyword{melt}
