% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_spectra.R
\name{normalize_spectra}
\alias{normalize_spectra}
\title{Average Spectra
Takes a tibble of grouped NORMALIZED spectrum data (multiple spectra per group)
returns ion-wise averaged spectra with averaged continuous statistics and nested categorical identifiers.
Some reserved column names get special treatment; this should become more generic in the future.}
\usage{
normalize_spectra(spectra)
}
\arguments{
\item{spectra}{Tibble featuring \code{mz}/\code{wl} and \code{intensity}. Should be grouped by spectrum.}
}
\description{
Average Spectra
Takes a tibble of grouped NORMALIZED spectrum data (multiple spectra per group)
returns ion-wise averaged spectra with averaged continuous statistics and nested categorical identifiers.
Some reserved column names get special treatment; this should become more generic in the future.
}
\examples{
spectra_norm <- spectra \%>\%
  group_by(scan) \%>\%
  normalize_spectra()

}
\keyword{average}
\keyword{consensus}
\keyword{spectra}
