% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaklist2spec.R
\name{peaklist2spec}
\alias{peaklist2spec}
\title{Convert (tbl) Peaklist to Spectrum}
\usage{
peaklist2spec(peaklist)
}
\arguments{
\item{peaklist}{The tibble in question}
}
\description{
Take tibble with mz and intensity columns (rt column optional).
Return a Spectrum object.
}
\examples{
my_spectrum <- peaklist2spec(my_peaklist)

}
\keyword{convert}
\keyword{encode}
\keyword{peaklist}
\keyword{spectrum}
