% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xic.R
\name{plot_xic}
\alias{plot_xic}
\title{Plot Extracted Ion Chromatogram}
\usage{
plot_xic(coords, alpha = 0.4)
}
\arguments{
\item{coords}{rt and intensity coords}
}
\description{
Take a table of rt and intensity coordinates, return a list of gg objects
representing a chromatogram with baseline-corrected area filled.
}
\examples{
xic_gg <- plot_xic(chromdata \%>\% filter(roi == 2))

}
\keyword{cosine}
\keyword{match}
\keyword{plot}
\keyword{spectra}
