% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tidymass.R
\name{read_tidymass}
\alias{read_tidymass}
\title{Read Tidy Mass Chromatogram from File}
\usage{
read_tidymass(file)
}
\arguments{
\item{file}{A tibble with columns scan or scan_num, mz, and intensity}
}
\description{
Read Tidy Mass Chromatogram from File
}
\examples{
ions \%>\% tbl2exp(file = "mychrom.mzxml")

}
\keyword{mz}
\keyword{tibble}
\keyword{write}
