% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reload_tidychrom}
\alias{reload_tidychrom}
\title{Author's utility functions, to be loaded with \code{tidychrom} namespace}
\usage{
reload_tidychrom()
}
\description{
Author's utility functions, to be loaded with \code{tidychrom} namespace
}
