% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_signals.R
\name{separate_signals}
\alias{separate_signals}
\title{Separate Spectral Signals
Takes a grouped tibble of spectral data, returns a summary of the input
with an x-coordinate that is diagnostic of each of the input spectra.}
\usage{
separate_signals(
  spectra,
  thres_ortho = 1,
  thres_intensity = 0.5,
  x = "mz",
  bin,
  cores = 1
)
}
\arguments{
\item{spectra}{Grouped tibble \code{mz}/\code{wl} and \code{intensity}. Should be grouped by scan.}

\item{thres_ortho}{Minimum orthogonality that a signal is allowed to have (1 is max)}

\item{thres_intensity}{Minimum relative intensity that a signal is allowed to have (0-1)}

\item{x}{Name of x-axis variable (\code{mz} or \code{wl}, etc.)}
}
\value{
Summary of input tibble with diagnostic x-coordinate
}
\description{
Separate Spectral Signals
Takes a grouped tibble of spectral data, returns a summary of the input
with an x-coordinate that is diagnostic of each of the input spectra.
}
\examples{
signal_ions <- conflicting_spectra \%>\%
  group_by(scan) \%>\%
  separate signals()

}
\keyword{cluster}
\keyword{cosine}
\keyword{spectra}
