% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_runs.R
\name{subtract_runs}
\alias{subtract_runs}
\title{Subtract Chromatographic Runs}
\usage{
subtract_runs(exp, samp_index_A, samp_index_B, bin)
}
\arguments{
\item{exp}{MSNExp object; both runs need to be in the experiment.}

\item{samp_index_A}{Index of first sample, where result = A - B.}

\item{samp_index_B}{Index of second sample (e.g. blank), where result = A - B.}

\item{bin}{Bin spectra to this m/z tolerance. If missing, do not bin at all.}
}
\description{
Take an MSNExp and two sample indices. Subtract the second index from the first,
save the resultant CDF, and return that filename.
}
\examples{
exp_blanked <- subtract_blank_exp(exp, blank_i = c(1,2))

}
\keyword{baseline}
\keyword{blank}
\keyword{spectra}
\keyword{spectrum}
\keyword{subtract}
