% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_spectra.R
\name{subtract_spectra}
\alias{subtract_spectra}
\title{Subtract Spectra}
\usage{
subtract_spectra(spec_samp, spec_bl, bin)
}
\arguments{
\item{spec_samp}{The spectrum with material in it}

\item{bin}{Bin spectra to this m/z tolerance. If missing, do not bin at all.}

\item{spec_blank}{The blank spectrum}
}
\description{
Take two MSnbase Spectrum objects and subtract one from the other, no scaling or anything.
Return another Spectrum. Use for subtracting a blank baseline.
}
\examples{
spec_corrected <- subtract_spectrum(specSamp, specBL)

}
\keyword{baseline}
\keyword{blank}
\keyword{spectra}
\keyword{spectrum}
\keyword{subtract}
