% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2exp.R
\name{tbl2exp}
\alias{tbl2exp}
\title{Convert Tibble to MSnExp}
\usage{
tbl2exp(ions)
}
\arguments{
\item{ions}{A tibble with columns feature, mz, and intensity}
}
\description{
Pack a tibble into an MSNExp object.
}
\examples{
expt <- tbl2exp(ions)

}
\keyword{convert}
\keyword{experiment}
