% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tidymass.R
\name{write_tidymass}
\alias{write_tidymass}
\title{Write Tidy Mass Chromatogram to File}
\usage{
write_tidymass(ions, file)
}
\arguments{
\item{ions}{A tibble with columns scan or scan_num, mz, and intensity}
}
\description{
Write Tidy Mass Chromatogram to File
}
\examples{
ions \%>\% tbl2exp(file = "mychrom.mzxml")

}
\keyword{mz}
\keyword{tibble}
\keyword{write}
