% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_from_db.R
\name{annot_from_db}
\alias{annot_from_db}
\title{Annotate from Database}
\usage{
annot_from_db(spectra, dir_db, bin, fname_regex = "*.mzXML", cores = 1)
}
\arguments{
\item{spectra}{Grouped (e.g. by \code{scan} tibble with columns \code{mz}, \code{intensity}.}

\item{dir_db}{Directory containing database spectrum files in an mzR-readable format (mzXML, mzML, etc.)}

\item{fname_regex}{Regex string that can be used to prefilter database files (generally the suffix.)}
}
\value{
A summary of the input tibble with additional columns:
\code{file} (basename of the best-matching database spectrum file)
\code{cos} (cosine matchup score for that file)
}
\description{
Simple iterative database search to annotate spectra in a tibble
}
\examples{
# effectively add an annotation column to peaks_matched:
peaks_matched <- annot_from_db(peaks_matched, "~/msdb/", "methyl")
}
\keyword{annot}
\keyword{annotate}
\keyword{database}
\keyword{db}
