% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve (AUC)}
\usage{
auc(coords, rt_peak, plot_intb = T)
}
\arguments{
\item{coords}{Tibble with columns \code{rt} and \code{intensity}}

\item{rt_peak}{Retention time of the focal peak in units matching coords}

\item{plot_intb}{Logical specifying whether to return ggplot objects}
}
\value{
a named list of:
\code{rtmin}, \code{rtmax},
\code{into} (integral overall),
\code{into}, (integral overall),
\code{intb}, (integral baseline-corrected),
\code{gg}, a vector of gg objects depicting the integrated peak
}
\description{
take molten chrom coords and an RT
return info about the peak, including
i.e. integrate a single peak
}
\examples{

peaks_matched <- annot_from_db(peaks_matched, "~/msdb/", "methyl")
}
\keyword{area}
\keyword{curve}
\keyword{integral}
\keyword{peak}
