% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_spectra.R
\name{average_spectra}
\alias{average_spectra}
\title{Average Spectra
Takes a tibble of grouped NORMALIZED spectrum data (multiple spectra per group)
returns ion-wise averaged spectra with averaged continuous statistics and nested categorical identifiers.
Some reserved column names get special treatment; this should become more generic in the future.}
\usage{
average_spectra(spectra, x = "mz")
}
\arguments{
\item{spectra}{Grouped tibble with \code{mz}/\code{wl} and \code{intensity}. Should be grouped by clusters to be averaged.}

\item{x}{Name of x-axis variable (\code{mz} or \code{wl}, etc.)}
}
\description{
Average Spectra
Takes a tibble of grouped NORMALIZED spectrum data (multiple spectra per group)
returns ion-wise averaged spectra with averaged continuous statistics and nested categorical identifiers.
Some reserved column names get special treatment; this should become more generic in the future.
}
\examples{
spectra_cons <- spectra_clust \%>\%
  group_by(clust_id) \%>\%
  average_spectra()

}
\keyword{average}
\keyword{consensus}
\keyword{spectra}
