% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ldrs.r
\name{calc_intb_max}
\alias{calc_intb_max}
\title{Calculate Linear Dynamic Ranges}
\usage{
calc_intb_max(peak_group, key, rsq = 0.95, min_dils = 3)
}
\arguments{
\item{rsq}{minimum R^2 at which signal is considered linear}

\item{peaks_matched}{Tibble with columns \code{rt_std} (or other compound identifier), \code{dil}, and \code{intb}}

\item{min_concs}{minimum number of dilutions that must be present for a compound}
}
\value{
tibble with compound ID, max. linear peak area, and R^2
}
\description{
Calculate Linear Dynamic Ranges
}
\examples{
# effectively add an annotation column to peaks_matched:
std_ldrs <- calc_ldrs(peaks_matched, 0.98, 4)
}
\keyword{dynamic}
\keyword{ldr}
\keyword{linear}
\keyword{range}
