% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_by_roi_scan.R
\name{cosine_by_roi_scan}
\alias{cosine_by_roi_scan}
\title{Compare Spectra by ROI and Scan}
\usage{
cosine_by_roi_scan(spectra_candidate, spectra_master, viz = F)
}
\arguments{
\item{spectra_candidate}{Tibble containing full scans and an roi column. Can be multiple scans/ROI.}

\item{spectra_master}{Tibble containing standard spectra, also identified with an roi number. Should be only 1 scan/ROI.}
}
\description{
Take candidate spectra, return a summary of the best matches against a set of master spectra.
}
\examples{
scan_cosines <- cosine_by_roi_scan(spectra_candidate, spectra_master)

}
\keyword{cosine}
\keyword{dot}
\keyword{product}
\keyword{spectra}
