% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_back2back.R
\name{plot_back2back}
\alias{plot_back2back}
\title{Plot Spectra Back-to-Back}
\usage{
plot_back2back(ions, pal = c("blue", "red"), x = "mz", label_pctile = 5)
}
\arguments{
\item{ions}{Grouped tibble containing mz and intensity.}

\item{colors}{Character vector of colors to plot the spectra in}
}
\description{
Take candidate spectra, return a summary of the best matches against a set of master spectra.
}
\examples{
scan_cosines <- cosine_by_roi_scan(spectra_candidate, spectra_master)

}
\keyword{cosine}
\keyword{match}
\keyword{plot}
\keyword{spectra}
