% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tidymass.R
\name{read_tidymass}
\alias{read_tidymass}
\title{Read Tidy Mass Chromatogram from File}
\usage{
read_tidymass(file)
}
\arguments{
\item{file}{Path to an mzR-compatible MS1 file}
}
\value{
A tibble with columns \code{scan}, \code{rt}, \code{mz}, and \code{intensity}
}
\description{
Read Tidy Mass Chromatogram from File
}
\examples{
ions <- read_tidymass(file = "mychrom.mzxml")

}
\keyword{mz}
\keyword{tibble}
\keyword{write}
